+++ src/App.tsx
@@ -24,8 +24,8 @@ import DefaultLoader from 'src/components/DefaultLoader';
 import { request, response } from 'src/store/reducers/resources';
 import Footer from 'src/components/Footer';
 
-const ListLinodes = DefaultLoader({
-  loader: () => import('src/features/linodes/ListLinodes'),
+const LinodesLanding = DefaultLoader({
+  loader: () => import('src/features/linodes/LinodesLanding'),
 });
 
 const theme = createMuiTheme(LinodeTheme as Linode.TodoAny);
@@ -139,7 +139,7 @@ export class App extends React.Component<FinalProps, State> {
               <div className={classes.wrapper}>
                 <Switch>
                   <TempRoute exact path="/dashboard" render={() => 'Dashboard'} />
-                  <Route exact path="/linodes" component={ListLinodes} />
+                  <Route exact path="/linodes" component={LinodesLanding} />
                   <TempRoute exact path="/volumes" render={() => 'Volumes'} />
                   <TempRoute exact path="/nodebalancers" render={() => 'NodeBalancers'} />
                   <TempRoute exact path="/domains" render={() => 'Domains'} />
