+++ src/features/linodes/LinodesLanding/LinodesLanding.test.tsx
@@ -2,7 +2,7 @@ import * as React from 'react';
 import { mount } from 'enzyme';
 import { StaticRouter } from 'react-router-dom';
 import { PromiseLoaderResponse } from 'src/components/PromiseLoader';
-import { RoutedListLinodes } from './index';
+import { RoutedListLinodes } from './LinodesLanding';
 import {
   linodes as mockLinodes,
   images as mockImages,
