+++ src/features/linodes/LinodesLanding/index.tsx
@@ -1,0 +1,2 @@
+import LinodesLanding from './LinodesLanding';
+export default LinodesLanding;
